requirements.txt
Django>=4.2
djangorestframework
python-dotenv
djangorestframework-simplejwt
requests
django-cors-headers
